--- config.ts
@@ -1,11 +1,0 @@
-import { browser, Config } from 'protractor' ;
-
-export const config: Config = {
-	framework: 'jasmine' ,
-	seleniumAddress: 'http://localhost:4444/wd/hub' ,
-	specs: [ '../test/Google.spec.js' ],
-	noGlobals: true ,
-	onPrepare: () => {
-		browser.ignoreSynchronization = true ;
-	}
-}
