+++ package.json
@@ -4,10 +4,10 @@
   "description": "This is a Workshop about Protractor",
   "main": "index.js",
   "scripts": {
-    "clean": "rm -rf dist" ,
-    "prebuild": "npm run clean" ,
-    "build": "tsc" ,
-    "test": "npm run build && protractor dist/config.js"
+    "clean": "rm -rf dist",
+    "prebuild": "npm run clean",
+    "build": "tsc",
+    "test": "npm run build && protractor dist/protractor/config.js"
   },
   "repository": {
     "type": "git",
@@ -26,6 +26,7 @@
   },
   "homepage": "https://github.com/JDGonzal/protractor-workshop-2017#readme",
   "dependencies": {
+    "jasmine-spec-reporter": "^4.2.1",
     "protractor": "^5.1.2"
   },
   "devDependencies": {
