+++ protractor/config.ts
@@ -1,0 +1,13 @@
+import { browser, Config } from 'protractor' ;
+import { reporter } from './helpers/reporter';
+
+export const config: Config = {
+  framework: 'jasmine' ,
+  seleniumAddress: 'http://localhost:4444/wd/hub' ,
+  specs: [ '../test/Google.spec.js' ],
+  noGlobals: true ,
+  onPrepare: () => {
+    browser.ignoreSynchronization = true ;
+    reporter();
+  }
+}
