+++ protractor/helpers/reporter.ts
@@ -1,0 +1,7 @@
+import { DisplayProcessor, SpecReporter } from "jasmine-spec-reporter" ;
+
+export let reporter = () => {
+  jasmine.getEnv().addReporter( new SpecReporter({
+    customProcessors: [DisplayProcessor],
+  }));
+};
